cometbft gordian integration
