+++ simapp/v2/simdv2/cmd/root_di.go
@@ -19,7 +19,7 @@ import (
 	"github.com/rollchains/gordian/gcosmos/gserver"
 	"github.com/rollchains/gordian/gcosmos/gtx"
 
-	// "cosmossdk.io/server/v2/cometbft" // TODO: spawntag:cometbft
+	"cosmossdk.io/server/v2/cometbft"
 	"cosmossdk.io/simapp/v2"
 	"cosmossdk.io/x/auth/tx"
 	authtxconfig "cosmossdk.io/x/auth/tx/config"
@@ -41,32 +41,40 @@ func NewRootCmdWithServer[T transaction.Tx](
 
 // NewRootCmd creates a new root command for simd. It is called once in the main function.
 func NewRootCmd[T transaction.Tx]() *cobra.Command {
-	// TODO: spawntag:cometbft multiline
-	// return NewRootCmdWithServer(func(cc client.Context) serverv2.ServerComponent[T] {
-	// 	return cometbft.New[T](
-	// 		&genericTxDecoder[T]{cc.TxConfig},
-	// 		cometbft.DefaultServerOptions[T](),
-	// 	)
-	// })
+	var cmd *cobra.Command
 
-	return NewRootCmdWithServer(func(cc client.Context) serverv2.ServerComponent[transaction.Tx] {
-		dc := server.NewDefaultContext() // "github.com/cosmos/cosmos-sdk/server"
+	// <spawntag:cometbft
+	cmd = NewRootCmdWithServer(func(cc client.Context) serverv2.ServerComponent[T] {
+		return cometbft.New[T](
+			&genericTxDecoder[T]{cc.TxConfig},
+			cometbft.DefaultServerOptions[T](),
+		)
+	})
+	// cometbft:cometbft>
 
-		ctx := context.WithValue(context.Background(), client.ClientContextKey, &client.Context{
+	// <spawntag:gordian
+	cmd = NewRootCmdWithServer(func(cc client.Context) serverv2.ServerComponent[transaction.Tx] {
+		dc := server.NewDefaultContext()
+
+		var ctx context.Context = context.Background()
+		ctx = context.WithValue(ctx, client.ClientContextKey, client.Context{
 			ChainID: "gcosmos",
 			HomeDir: "/home/reece/.simappv2",
 		})
-		ctx = context.WithValue(ctx, server.ServerContextKey, dc)
+		// ctx = context.WithValue(ctx, server.ServerContextKey, server.NewDefaultContext())
 
 		log := slog.New(slog.NewJSONHandler(os.Stdout, nil))
 
 		codec := gtx.NewTxDecoder(cc.TxConfig)
-		c, err := gserver.NewComponent(ctx, log, codec, cc.Codec)
+		c, err := gserver.NewComponent(ctx, dc, log, codec, cc.Codec)
 		if err != nil {
 			panic(err)
 		}
 		return c
 	})
+	// spawntag:gordian>
+
+	return cmd
 }
 
 func newRootCmd[T transaction.Tx](
